+++ meta-memfault/recipes-memfault/memfaultd/files/memfaultd/src/bin/mfw.rs
@@ -1,0 +1,11 @@
+//
+// Copyright (c) Memfault, Inc.
+// See License.txt for details
+use memfaultd::cli;
+
+/// mfw is an alias to the main function in cli.rs
+///
+/// For further details, see comments on memfaultd.rs
+fn main() {
+    cli::main()
+}