+++ meta-memfault/recipes-memfault/memfaultd/files/memfaultd/src/metrics/system_metrics/mod.rs
@@ -1,77 +1,180 @@
 //
 // Copyright (c) Memfault, Inc.
 // See License.txt for details
-use std::sync::{Arc, Mutex};
-use std::thread::{self, spawn};
-use std::time::Duration;
+use std::{
+    collections::HashSet,
+    sync::{Arc, Mutex},
+    thread::sleep,
+    time::{Duration, Instant},
+};
 
 use eyre::Result;
 use log::warn;
 
-use crate::metrics::MetricReportManager;
+use crate::{
+    config::SystemMetricConfig,
+    metrics::{KeyedMetricReading, MetricReportManager},
+    util::system::{bytes_per_page, clock_ticks_per_second},
+};
+
 mod cpu;
 use crate::metrics::system_metrics::cpu::{CpuMetricCollector, CPU_METRIC_NAMESPACE};
+
+mod thermal;
+use crate::metrics::system_metrics::thermal::{ThermalMetricsCollector, THERMAL_METRIC_NAMESPACE};
 
 mod memory;
 use crate::metrics::system_metrics::memory::{MemoryMetricsCollector, MEMORY_METRIC_NAMESPACE};
 
-pub const BUILTIN_SYSTEM_METRIC_NAMESPACES: &[&str; 2] =
-    &[CPU_METRIC_NAMESPACE, MEMORY_METRIC_NAMESPACE];
+mod network_interfaces;
+use crate::metrics::system_metrics::network_interfaces::{
+    NetworkInterfaceMetricCollector, NetworkInterfaceMetricsConfig,
+    NETWORK_INTERFACE_METRIC_NAMESPACE,
+};
 
-pub struct SystemMetricsCollector {}
+mod processes;
+use processes::{ProcessMetricsCollector, PROCESSES_METRIC_NAMESPACE};
+
+mod disk_space;
+use disk_space::{
+    DiskSpaceMetricCollector, DiskSpaceMetricsConfig, NixStatvfs, DISKSPACE_METRIC_NAMESPACE,
+    DISKSPACE_METRIC_NAMESPACE_LEGACY,
+};
+
+use self::processes::ProcessMetricsConfig;
+
+pub const BUILTIN_SYSTEM_METRIC_NAMESPACES: &[&str; 7] = &[
+    CPU_METRIC_NAMESPACE,
+    MEMORY_METRIC_NAMESPACE,
+    THERMAL_METRIC_NAMESPACE,
+    NETWORK_INTERFACE_METRIC_NAMESPACE,
+    PROCESSES_METRIC_NAMESPACE,
+    DISKSPACE_METRIC_NAMESPACE,
+    // Include in list of namespaces so that
+    // legacy collectd from the "df" plugin
+    // are still filtered out
+    DISKSPACE_METRIC_NAMESPACE_LEGACY,
+];
+
+pub trait SystemMetricFamilyCollector {
+    fn collect_metrics(&mut self) -> Result<Vec<KeyedMetricReading>>;
+    fn family_name(&self) -> &'static str;
+}
+
+pub struct SystemMetricsCollector {
+    metric_family_collectors: Vec<Box<dyn SystemMetricFamilyCollector>>,
+}
 
 impl SystemMetricsCollector {
-    pub fn new() -> Self {
-        Self {}
+    pub fn new(system_metric_config: SystemMetricConfig) -> Self {
+        // CPU, Memory, and Thermal metrics are captured by default
+        let mut metric_family_collectors: Vec<Box<dyn SystemMetricFamilyCollector>> = vec![
+            Box::new(CpuMetricCollector::new()),
+            Box::new(MemoryMetricsCollector::new()),
+            Box::new(ThermalMetricsCollector::new()),
+        ];
+
+        // Check if process metrics have been manually configured
+        match system_metric_config.processes {
+            Some(processes) if !processes.is_empty() => {
+                metric_family_collectors.push(Box::new(ProcessMetricsCollector::<Instant>::new(
+                    ProcessMetricsConfig::Processes(processes),
+                    clock_ticks_per_second() as f64 / 1000.0,
+                    bytes_per_page() as f64,
+                )))
+            }
+            // Monitoring no processes means this collector is disabled
+            Some(_empty_set) => {}
+            None => {
+                metric_family_collectors.push(Box::new(ProcessMetricsCollector::<Instant>::new(
+                    ProcessMetricsConfig::Auto,
+                    clock_ticks_per_second() as f64 / 1000.0,
+                    bytes_per_page() as f64,
+                )))
+            }
+        };
+
+        // Check if disk space metrics have been manually configured
+        match system_metric_config.disk_space {
+            Some(disks) if !disks.is_empty() => {
+                metric_family_collectors.push(Box::new(DiskSpaceMetricCollector::new(
+                    NixStatvfs::new(),
+                    DiskSpaceMetricsConfig::Disks(disks),
+                )))
+            }
+            // Monitoring no disks means this collector is disabled
+            Some(_empty_set) => {}
+            None => metric_family_collectors.push(Box::new(DiskSpaceMetricCollector::new(
+                NixStatvfs::new(),
+                DiskSpaceMetricsConfig::Auto,
+            ))),
+        };
+
+        // Check if network interface metrics have been manually configured
+        match system_metric_config.network_interfaces {
+            Some(interfaces) if !interfaces.is_empty() => metric_family_collectors.push(Box::new(
+                NetworkInterfaceMetricCollector::<Instant>::new(
+                    NetworkInterfaceMetricsConfig::Interfaces(interfaces),
+                ),
+            )),
+            // Monitoring no interfaces means this collector is disabled
+            Some(_empty_set) => {}
+            None => metric_family_collectors.push(Box::new(NetworkInterfaceMetricCollector::<
+                Instant,
+            >::new(
+                NetworkInterfaceMetricsConfig::Auto
+            ))),
+        };
+
+        Self {
+            metric_family_collectors,
+        }
     }
 
-    pub fn start(
-        &self,
+    pub fn run(
+        &mut self,
         metric_poll_duration: Duration,
         metric_report_manager: Arc<Mutex<MetricReportManager>>,
-    ) -> Result<()> {
-        let mut cpu_metric_collector = CpuMetricCollector::new();
-
-        spawn(move || loop {
-            thread::sleep(metric_poll_duration);
-
-            match cpu_metric_collector.get_cpu_metrics() {
-                Ok(cpu_metrics) => {
-                    for metric_reading in cpu_metrics {
-                        if let Err(e) = metric_report_manager
-                            .lock()
-                            .expect("Mutex poisoned")
-                            .add_metric(metric_reading)
-                        {
-                            warn!("Couldn't add CPU metric: {}", e)
+    ) {
+        loop {
+            for collector in self.metric_family_collectors.iter_mut() {
+                match collector.collect_metrics() {
+                    Ok(readings) => {
+                        for metric_reading in readings {
+                            if let Err(e) = metric_report_manager
+                                .lock()
+                                .expect("Mutex poisoned")
+                                .add_metric(metric_reading)
+                            {
+                                warn!(
+                                    "Couldn't add metric reading for family \"{}\": {:?}",
+                                    collector.family_name(),
+                                    e
+                                )
+                            }
                         }
                     }
+                    Err(e) => warn!(
+                        "Failed to collect readings for family \"{}\": {}",
+                        collector.family_name(),
+                        e
+                    ),
                 }
-                Err(e) => warn!("CPU metric collection failed: {}", e),
             }
 
-            match MemoryMetricsCollector::get_memory_metrics() {
-                Ok(memory_metric_readings) => {
-                    for metric_reading in memory_metric_readings {
-                        if let Err(e) = metric_report_manager
-                            .lock()
-                            .expect("Mutex poisoned")
-                            .add_metric(metric_reading)
-                        {
-                            warn!("Couldn't add memory metric reading: {:?}", e)
-                        }
-                    }
-                }
-                Err(e) => warn!("Memory metric collection failed: {:?}", e),
-            }
-        });
-
-        Ok(())
+            sleep(metric_poll_duration);
+        }
     }
 }
 
 impl Default for SystemMetricsCollector {
     fn default() -> Self {
-        Self::new()
+        Self::new(SystemMetricConfig {
+            enable: true,
+            poll_interval_seconds: Duration::from_secs(10),
+            processes: Some(HashSet::from_iter(["memfaultd".to_string()])),
+            disk_space: None,
+            network_interfaces: None,
+        })
     }
 }